finishes secret-handshake for elixir
This one took alot out of me, trying to keep it small was tough, but looking at the other solutions on Exercism, it looks like my footprint was pretty good. I think it was a good decision to filter with the charlist as opposed to using the Bitwise stuff. Overall happy with how this one turned out.